feat(filters): add DateRangeChips to 5 list pages + share component
Move DateRangeChips (ทั้งหมด / เดือนนี้ / เดือนที่แล้ว / ช่วงวันที่...) from
pages/other-income/components/ → components/ui/ so any page can import. Old
path keeps a re-export for backward compat. ThaiDateInput now forwards a
`data-date-range-custom-start` attribute so the "ช่วงวันที่..." chip can
focus the start input.

Pages getting the chips:
- /audit-logs        — dateFrom/dateTo
- /finance/general-ledger
- /reports           — entity profit section
- /contracts         — inside the collapsible date filter
- /finance-portfolio — new filter (also adds startDate/endDate to API,
  filters Contract.createdAt)
