+++ apps/web/src/components/ui/DateRangeChips.tsx
@@ -1,0 +1,147 @@
+import { Calendar } from 'lucide-react';
+import { THAI_MONTHS_FULL, THAI_MONTHS_SHORT } from '@/lib/date';
+
+interface DateRangeChipsProps {
+  startDate: string; // YYYY-MM-DD or ''
+  endDate: string;   // YYYY-MM-DD or ''
+  onChange: (next: { startDate: string; endDate: string }) => void;
+  /** When false, hide the "ทั้งหมด" preset (use on pages where empty range is invalid). */
+  showAllChip?: boolean;
+}
+
+function toIsoDate(d: Date): string {
+  const y = d.getFullYear();
+  const m = String(d.getMonth() + 1).padStart(2, '0');
+  const day = String(d.getDate()).padStart(2, '0');
+  return `${y}-${m}-${day}`;
+}
+
+function thisMonthRange(today: Date): { startDate: string; endDate: string } {
+  const first = new Date(today.getFullYear(), today.getMonth(), 1);
+  return { startDate: toIsoDate(first), endDate: toIsoDate(today) };
+}
+
+function lastMonthRange(today: Date): { startDate: string; endDate: string } {
+  const first = new Date(today.getFullYear(), today.getMonth() - 1, 1);
+  const last = new Date(today.getFullYear(), today.getMonth(), 0);
+  return { startDate: toIsoDate(first), endDate: toIsoDate(last) };
+}
+
+function isSameRange(
+  a: { startDate: string; endDate: string },
+  b: { startDate: string; endDate: string },
+): boolean {
+  return a.startDate === b.startDate && a.endDate === b.endDate;
+}
+
+function formatRangeLabel(startDate: string, endDate: string): string {
+  if (!startDate && !endDate) return 'ทั้งหมด';
+  if (!startDate || !endDate) return `${startDate || endDate}`;
+
+  const start = new Date(startDate + 'T00:00:00');
+  const end = new Date(endDate + 'T00:00:00');
+  const beYear = start.getFullYear() + 543;
+  const startMonthIdx = start.getMonth();
+  const endMonthIdx = end.getMonth();
+  const startDay = start.getDate();
+  const endDay = end.getDate();
+  const lastDayOfStartMonth = new Date(start.getFullYear(), startMonthIdx + 1, 0).getDate();
+
+  const sameYear = start.getFullYear() === end.getFullYear();
+  const sameMonth = sameYear && startMonthIdx === endMonthIdx;
+
+  if (sameMonth && startDay === 1 && endDay === lastDayOfStartMonth) {
+    return `${THAI_MONTHS_FULL[startMonthIdx]} ${beYear}`;
+  }
+  if (sameMonth) {
+    const sd = String(startDay).padStart(2, '0');
+    const ed = String(endDay).padStart(2, '0');
+    const mm = String(startMonthIdx + 1).padStart(2, '0');
+    return `${THAI_MONTHS_FULL[startMonthIdx]} ${beYear} (${sd}/${mm} - ${ed}/${mm})`;
+  }
+  const endBeYear = end.getFullYear() + 543;
+  if (sameYear) {
+    return `${startDay} ${THAI_MONTHS_SHORT[startMonthIdx]} - ${endDay} ${THAI_MONTHS_SHORT[endMonthIdx]} ${endBeYear}`;
+  }
+  return `${startDay} ${THAI_MONTHS_SHORT[startMonthIdx]} ${beYear} - ${endDay} ${THAI_MONTHS_SHORT[endMonthIdx]} ${endBeYear}`;
+}
+
+export function DateRangeChips({ startDate, endDate, onChange, showAllChip = true }: DateRangeChipsProps) {
+  const today = new Date();
+  const current = { startDate, endDate };
+  const presets = {
+    all: { startDate: '', endDate: '' },
+    thisMonth: thisMonthRange(today),
+    lastMonth: lastMonthRange(today),
+  };
+
+  const isAll = isSameRange(current, presets.all);
+  const isThisMonth = isSameRange(current, presets.thisMonth);
+  const isLastMonth = isSameRange(current, presets.lastMonth);
+  const isCustom = !isAll && !isThisMonth && !isLastMonth;
+
+  const chipClass = (active: boolean) =>
+    `px-3 py-1.5 text-xs font-medium rounded-md border transition-colors ${
+      active
+        ? 'bg-primary text-primary-foreground border-primary'
+        : 'bg-background text-foreground border-border hover:bg-accent'
+    }`;
+
+  return (
+    <div className="flex items-center justify-between flex-wrap gap-2 w-full">
+      <div role="radiogroup" aria-label="ช่วงวันที่" className="flex flex-wrap gap-2">
+        {showAllChip && (
+          <button
+            type="button"
+            role="radio"
+            aria-checked={isAll}
+            className={chipClass(isAll)}
+            onClick={() => onChange(presets.all)}
+          >
+            ทั้งหมด
+          </button>
+        )}
+        <button
+          type="button"
+          role="radio"
+          aria-checked={isThisMonth}
+          className={chipClass(isThisMonth)}
+          onClick={() => onChange(presets.thisMonth)}
+        >
+          เดือนนี้
+        </button>
+        <button
+          type="button"
+          role="radio"
+          aria-checked={isLastMonth}
+          className={chipClass(isLastMonth)}
+          onClick={() => onChange(presets.lastMonth)}
+        >
+          เดือนที่แล้ว
+        </button>
+        <button
+          type="button"
+          role="radio"
+          aria-checked={isCustom}
+          aria-label="ช่วงวันที่ กำหนดเอง"
+          className={chipClass(isCustom)}
+          onClick={() => {
+            const customInput = document.querySelector<HTMLInputElement>(
+              'input[data-date-range-custom-start="true"]',
+            );
+            customInput?.focus();
+          }}
+        >
+          ช่วงวันที่...
+        </button>
+      </div>
+      <div
+        data-testid="date-range-label"
+        className="inline-flex items-center gap-1.5 text-xs text-muted-foreground"
+      >
+        <Calendar size={13} />
+        {formatRangeLabel(startDate, endDate)}
+      </div>
+    </div>
+  );
+}
